Rename GlobalDiscovery and LocalDiscovery into MultiNodeDiscovery and LocalhostDiscovery and fix some names after applying review suggestion
